+++ src/shared/nested_apply.rs
@@ -12,10 +12,18 @@ use crate::{
     variable_id::{VarIdPacked16, VarIdPacked32, VarIdPacked64, VarIdPackedAny, VariableId},
 };
 
-use super::{bdd::Bdd, manager::BddManager};
+use super::{manager::BddManager, Bdd};
 
 impl BddManager {
-    fn nested_apply<TOuterOp: BooleanOperator, TInnerOp: BooleanOperator>(
+    /// Applies the `outer_op` to the [`Bdd`]s. On each node of the resulting `Bdd`,
+    /// if its variable is in `variables`, the node is replaced with the result
+    /// of applying `inner_op` to its low and high children.
+    ///
+    /// This function is useful for implementing combinations of applying binary
+    /// operators and quantification. For example, if `outer_op` is [`boolean_operators::And`]
+    /// and `inner_op` is [`boolean_operators::Or`], this combination corresponds to the
+    /// "relational product" operation.
+    pub fn nested_apply<TOuterOp: BooleanOperator, TInnerOp: BooleanOperator>(
         &mut self,
         left: &Bdd,
         right: &Bdd,
@@ -74,17 +82,17 @@ impl BddManager {
         bdd
     }
 
-    /// Eliminates the given `variables` using existential quantification.
+    /// Calculates a [`Bdd`] with the given `variables` eliminated using existential quantification.
     pub fn exists(&mut self, bdd: &Bdd, variables: &[VariableId]) -> Bdd {
         self.binary_op_with_exists(bdd, bdd, boolean_operators::And, variables)
     }
 
-    /// Eliminates the given `variables` using universal quantification.
+    /// Calculates a [`Bdd`] with the given `variables` eliminated using universal quantification.
     pub fn for_all(&mut self, bdd: &Bdd, variables: &[VariableId]) -> Bdd {
         self.binary_op_with_for_all(bdd, bdd, boolean_operators::And, variables)
     }
 
-    /// Applies a binary operator to the BDDs and eliminates the given `variables` using existential
+    /// Applies the `operator` to the [`Bdd`]s and eliminates the given `variables` using existential
     /// quantification from the result.
     pub fn binary_op_with_exists<TBooleanOp: BooleanOperator>(
         &mut self,
@@ -96,7 +104,7 @@ impl BddManager {
         self.nested_apply(left, right, operator, boolean_operators::Or, variables)
     }
 
-    /// Applies a binary operator to the BDDs and eliminates the given `variables` using universal
+    /// Applies the `operator` to the [`Bdd`]s and eliminates the given `variables` using universal
     /// quantification from the result.
     pub fn binary_op_with_for_all<TBooleanOp: BooleanOperator>(
         &mut self,
@@ -420,7 +428,7 @@ fn nested_apply_64_bit<TOuterOp: BooleanOperator, TInnerOp: BooleanOperator>(
 mod tests {
     use crate::{
         boolean_operators,
-        shared::{bdd::Bdd, manager::BddManager},
+        shared::{manager::BddManager, Bdd},
         variable_id::VariableId,
     };
